+++ lib/model.py
@@ -4,7 +4,7 @@ from keras.layers.advanced_activations import LeakyReLU
 from keras.layers.convolutional import Conv2D
 from keras.optimizers import Adam
 
-from .pixel_shuffler import PixelShuffler
+from .PixelShuffler import PixelShuffler
 
 optimizer = Adam(lr=5e-5, beta_1=0.5, beta_2=0.999)
