+++ lib/utils.py
@@ -24,23 +24,3 @@ def load_images(image_paths, convert=None):
             all_images = numpy.empty((len(image_paths), ) + image.shape, dtype=image.dtype)
         all_images[i] = image
     return all_images
-
-
-def get_transpose_axes(n):
-    if n % 2 == 0:
-        y_axes = list(range(1, n - 1, 2))
-        x_axes = list(range(0, n - 1, 2))
-    else:
-        y_axes = list(range(0, n - 1, 2))
-        x_axes = list(range(1, n - 1, 2))
-    return y_axes, x_axes, [n - 1]
-
-
-def stack_images(images):
-    images_shape = numpy.array(images.shape)
-    new_axes = get_transpose_axes(len(images_shape))
-    new_shape = [numpy.prod(images_shape[x]) for x in new_axes]
-    return numpy.transpose(
-        images,
-        axes=numpy.concatenate(new_axes)
-        ).reshape(new_shape)
