+++ plugins/Extract_Crop.py
@@ -1,0 +1,5 @@
+import cv2
+
+class Extract(object):
+    def extract(self, image, face, size):
+        return cv2.resize(face.image, size)
