+++ scripts/convert.py
@@ -1,9 +1,9 @@
 import cv2
 from lib.cli import DirectoryProcessor, FullPaths
 from pathlib import Path
-from lib.faces_process import convert_one_image
-from lib.faces_detect import crop_faces
+from lib.faces_detect import detect_faces
 
+from plugins.Convert_Adjust import Convert
 
 class ConvertImage(DirectoryProcessor):
     filename = ''
@@ -32,13 +32,14 @@ class ConvertImage(DirectoryProcessor):
 
     def process_image(self, filename):
         try:
+            converter = Convert(self.arguments.model_dir, self.arguments.swap_model)
             image = cv2.imread(filename)
-            for (idx, face) in enumerate(crop_faces(image)):
+            for (idx, face) in enumerate(detect_faces(image)):
                 if idx > 0 and self.arguments.verbose:
                     print('- Found more than one face!')
                     self.verify_output = True
 
-                new_face = convert_one_image(cv2.resize(face.image, (256, 256)), self.arguments.model_dir, self.arguments.swap_model)
+                new_face = converter.convert_one_image(cv2.resize(face.image, (256, 256)))
                 image[slice(face.y, face.y + face.h), slice(face.x, face.x + face.w)] = cv2.resize(new_face, (face.w, face.h))
                 self.faces_detected = self.faces_detected + 1
             output_file = self.output_dir / Path(filename).name
