+++ scripts/extract.py
@@ -1,7 +1,9 @@
 import cv2
-from lib.cli import DirectoryProcessor
-from pathlib import Path
 
+from pathlib import Path
+from lib.cli import DirectoryProcessor
+from lib.faces_detect import detect_faces
+from plugins.Extract_Crop import Extract
 
 class ExtractTrainingData(DirectoryProcessor):
     def create_parser(self, subparser, command, description):
@@ -13,8 +15,18 @@ class ExtractTrainingData(DirectoryProcessor):
             https://github.com/deepfakes/faceswap-playground"
         )
         
-    def process_face(self, face, index, filename):
-        resized_image = cv2.resize(face.image, (256, 256))
-        output_file = self.output_dir / Path(filename).stem
-        cv2.imwrite(str(output_file) + str(index) + Path(filename).suffix,
-                    resized_image)
+    def process_image(self, filename):
+        extractor = Extract()
+        try:
+            image = cv2.imread(filename)
+            for (idx, face) in enumerate(detect_faces(image)):
+                if idx > 0 and self.arguments.verbose:
+                    print('- Found more than one face!')
+                    self.verify_output = True
+
+                resized_image = extractor.extract(image, face, (256, 256))
+                output_file = self.output_dir / Path(filename).stem
+                cv2.imwrite(str(output_file) + str(idx) + Path(filename).suffix, resized_image)
+                self.faces_detected = self.faces_detected + 1
+        except Exception as e:
+            print('Failed to extract from image: {}. Reason: {}'.format(filename, e))
